refactor(core): split rx/nng runtime out of yijinjing/common.h
yijinjing/common.h was a god-header: every journal-core consumer also
dragged in rxcpp, nng and the full longfist type registry. Split it along
the core/runtime boundary:

  - new include/kungfu/yijinjing/rx.h carries the whole kungfu::rx event-flow
    layer (rxcpp operators, is_custom_event over AllTypesTags, holdon, the
    $/$$ subscribe helpers) moved verbatim; practice/hero.h includes it, which
    covers apprentice, wingchun and the node binding transitively
  - common.h keeps only core surface (KB/MB, yijinjing_error, resource,
    publisher/observer, locator/location, hana::to_json) and now includes the
    longfist/core.h schema leaf instead of the full longfist.h
  - the publisher interface no longer needs nng: new core-side constant
    PUBLISH_NONBLOCK replaces NNG_FLAG_NONBLOCK as the default publish flag,
    and nanomsg/socket.h static_asserts it equals NNG_FLAG_NONBLOCK so the
    mirrored value cannot drift
  - journal/frame.h temporarily includes longfist.h itself for the typed
    to_string AllTypes loop (to be removed when that moves out of core);
    assemble.cpp / replay_writer.cpp add the std headers they had been
    getting through the god-header

Pure refactor, no behaviour change. libkungfu builds green and the
fact-ledger slice run verifies write/reopen/checksum end to end.
