+++ framework/core/src/libkungfu/yijinjing/journal/assemble.cpp
@@ -11,6 +11,7 @@
 #include <kungfu/yijinjing/journal/assemble.h>
 #include <kungfu/yijinjing/journal/bus.h>
 #include <kungfu/yijinjing/time.h>
+#include <unordered_set>
 
 namespace kungfu::yijinjing::journal {
 using namespace longfist::enums;
